python/requirements.txt
pyserial==3.5
opencv-python==4.7.0.68
mediapipe==0.10.13
tensorflow 
numpy
Pillow==10.0.0  # For GUI image processing